[V14] Effect Stacking (#1667)
* Added the ability for effects to have stacks

* Fixed effect stacking

* Improved token overlay spacing

* Compendium updaetes

* Simplify effect click event (#1748)

* Fixed a bunch of deprecations

* Corrected AgileScout Beastform json data

* Updated TokenHUD to the new v14

* Removed DestroyOnEmpty from consumables

* Fixed so that tooltips don't get stuck (#1745)

* [Feature] TagTeam Partial Rendering (#1735)

* I done did it, I think

* Think I fixed the partial rendering bug for gm->player

* [V14] 1743 - Damage Update Error (#1746)

* Fixed DamageParts causing errors on update

* Fixed ActionBaseConfig error when no damage present on the action

* Fix removal of damage field

* Removed unneccessary default value function for parts

---------


* Simplify effect click event

---------


* Fixed stacking-value pointer event

* Set the stacking value in EffectsDisplay to be tabular-nums for monospacing

* Made baseEffect.stacking nullable instead of having an enabled property

* .

* Fixed so that actor._onUpdateDescantDocuments re-renders the EffectDisplay if effects were updated

---------

+++ module/applications/ui/effectsDisplay.mjs
@@ -49,11 +49,9 @@ export default class DhEffectsDisplay extends HandlebarsApplicationMixin(Applica
 
     _attachPartListeners(partId, htmlElement, options) {
         super._attachPartListeners(partId, htmlElement, options);
-
-        if (this.element) {
-            this.element.querySelectorAll('.effect-container a').forEach(element => {
-                element.addEventListener('contextmenu', this.removeEffect.bind(this));
-            });
+        for (const element of this.element?.querySelectorAll('.effect-container a') ?? []) {
+            element.addEventListener('click', e => this.#onClickEffect(e));
+            element.addEventListener('contextmenu', e => this.#onClickEffect(e, -1));
         }
     }
 
@@ -87,11 +85,21 @@ export default class DhEffectsDisplay extends HandlebarsApplicationMixin(Applica
         this.render();
     }
 
-    async removeEffect(event) {
+    async #onClickEffect(event, delta = 1) {
         const element = event.target.closest('.effect-container');
         const effects = DhEffectsDisplay.getTokenEffects();
         const effect = effects.find(x => x.id === element.dataset.effectId);
-        await effect.delete();
+        if (!effect || (delta >= 0 && !effect.system.stacking)) {
+            return;
+        }
+
+        const maxValue = effect.system.stacking?.max ?? Infinity;
+        const newValue = Math.clamp((effect.system.stacking?.value ?? 1) + delta, 0, maxValue);
+        if (newValue > 0) {
+            await effect.update({ 'system.stacking.value': newValue });
+        } else {
+            await effect.delete();
+        }
         this.render();
     }
